3 review
closeModalByOverlay - removed to modal.jsx
closeModalByEscape - removed to modal.jsx
closeModal - removed to modal.jsx

import of component App renamed to app/App
