+++ src/components/modal/modal.jsx
@@ -7,16 +7,34 @@ import { CloseIcon } from "@ya.praktikum/react-developer-burger-ui-components";
 const modalRoot = document.querySelector("#modal-root");
 
 export const Modal = (props) => {
+    const closeModalByEscape = (evt) => {
+        if (evt.keyCode === 27) {
+            props.setModal(props.ingredientModalState === false);
+        }
+    };
+
+    const closeModalByOverlay = (evt) => {
+        if (evt.type === "click") {
+            if (evt.target === evt.currentTarget) {
+                props.setModal(props.ingredientModalState === false);
+            }
+        }
+    };
+
+    const closeModal = () => {
+        props.setModal(props.ingredientModalState === false);
+    };
+
     useEffect(() => {
-        window.addEventListener("keyup", props.closeModalByEscape);
+        window.addEventListener("keyup", closeModalByEscape);
         return () => {
-            window.removeEventListener("keyup", props.closeModalByEscape);
+            window.removeEventListener("keyup", closeModalByEscape);
         };
     });
     return ReactDOM.createPortal(
-        <div className={props.modalState === props.state ? `${styles.overlay_deactive}` : `${styles.overlay}`} onClick={props.closeModalByOverlay}>
+        <div className={props.modalState === props.state ? `${styles.overlay_deactive}` : `${styles.overlay}`} onClick={closeModalByOverlay}>
             <div className={`${styles.contentBox} p-10`}>
-                <div className={`${styles.closeButton}`} onClick={props.closeModal}>
+                <div className={`${styles.closeButton}`} onClick={closeModal}>
                     <CloseIcon type="primary" />
                 </div>
                 <div className={`${styles.content}`}>{props.children}</div>
