+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./components/app/app"; //low registr
+import App from "./components/app/App";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
